# docs/openapi-v3.yaml
openapi: 3.0.3
info:
  title: Product Requirements Management API
  description: |
    Comprehensive API for managing product requirements through hierarchical structure of Epics → User Stories → Requirements. 
    Features include full-text search, comment system, relationship mapping, and configurable workflows.
  version: 1.0.0
  contact:
    name: API Support
    url: http://www.swagger.io/support
    email: support@swagger.io
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
servers:
  - url: http://localhost:8080
    description: Development server
  - url: https://api.requirements.example.com
    description: Production server
security:
  - BearerAuth: []
paths:
...
  /api/v1/search:
    get:
      tags: [Search]
      summary: Global search
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
          description: Search query
        - name: entity_types
          in: query
          schema:
            type: string
          description: Comma-separated entity types to search
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/OffsetParam'
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/search/suggestions:
    get:
      tags: [Search]
      summary: Get search suggestions
      parameters:
        - name: query
          in: query
          required: true
          schema:
            type: string
            minLength: 2
          description: Partial search query
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 10
          description: Maximum suggestions per category
      responses:
        '200':
          description: Search suggestions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchSuggestionsResponse'

  # Hierarchy endpoints
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authentication. Include as Authorization Bearer token header.
  parameters:
...
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 50
      description: Maximum number of results
    OffsetParam:
      name: offset
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
      description: Number of results to skip
...
  responses:
...
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    EntityType:
      type: string
      enum: [epic, user_story, acceptance_criteria, requirement]
...
    ErrorResponse:
      type: object
      description: Standard error response with code and message
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
...
    SearchResult:
      type: object
      required: [entity_type, entity_id, reference_id, title, rank]
      properties:
        entity_type:
          $ref: '#/components/schemas/EntityType'
        entity_id:
          type: string
          format: uuid
        reference_id:
          type: string
        title:
          type: string
        description:
          type: string
        highlight:
          type: string
        rank:
          type: number
    SearchResponse:
      type: object
      required: [results, total_count, query, entity_types, limit, offset]
      properties:
        results:
          type: array
          items:
            $ref: '#/components/schemas/SearchResult'
        total_count:
          type: integer
          minimum: 0
        query:
          type: string
        entity_types:
          type: array
          items:
            type: string
        limit:
          type: integer
          minimum: 1
          maximum: 100
        offset:
          type: integer
          minimum: 0
    SearchSuggestionsResponse:
      type: object
      required: [titles, reference_ids, statuses]
      properties:
        titles:
          type: array
          items:
            type: string
        reference_ids:
          type: array
          items:
            type: string
        statuses:
          type: array
          items:
            type: string
...